+++ database/migrations/2019_05_15_211333_create_orders_table.php
@@ -14,9 +14,9 @@ class CreateOrdersTable extends Migration
     public function up()
     {
         Schema::create('orders', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('no')->unique()->comment('订单流水号');
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->text('address')->comment('JSON 格式的收获地址');
             $table->decimal('total_amount', 10, 2)->comment('订单总金额');
