+++ database/migrations/2019_07_16_212955_orders_add_coupon_code_id.php
@@ -14,7 +14,7 @@ class OrdersAddCouponCodeId extends Migration
     public function up()
     {
         Schema::table('orders', function (Blueprint $table) {
-            $table->unsignedInteger('coupon_code_id')->nullable()->after('paid_at');
+            $table->unsignedBigInteger('coupon_code_id')->nullable()->after('paid_at');
             $table->foreign('coupon_code_id')->references('id')->on('coupon_codes')->onDelete('set null');
         });
     }
